+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 4.4.27 - 2020-11-03
+### Added
+* Added `!rule` command for quickly displaying a rule (closes [#200](https://github.com/cbrxyz/pi-bot/issues/200))
+
+### Removed
+* Removed `!invites` command, which led to nothing
+
 ## 4.4.26 - 2020-11-02
 ### Added
 * Commandinfo for each command can now support flags (closes [#29](https://github.com/cbrxyz/pi-bot/issues/29))